create a section heading component to harmonize section headings across the plateform
